+++ src/core/mixins/client-websockets.js
@@ -22,6 +22,61 @@ const ClientWebsockets = {
      * @private
      */
     'websocket:operation',
+
+    /**
+     * Websocket connection request is about to be made
+     * @event websocket:connecting
+     */
+    'websocket:connecting',
+
+    /**
+     * Websocket is being closed
+     *
+     * @event websocket:disconnecting
+     * @param {Layer.Core.Event} evt
+     * @param {String} evt.from  Describes where the call to scheduleReconnect originated from
+     * @param {String} evt.why   Provides detail on why the call was made from there
+     */
+    'websocket:disconnecting',
+
+    /**
+     * Websocket Event.replay is about to be issued
+     * @event websocket:replaying-events
+     * @param {Layer.Core.Event} evt
+     * @param {String} evt.from  Describes where the call to Event.replay originated from
+     * @param {String} evt.why   Provides detail on why the call was made from there
+     */
+    'websocket:replaying-events',
+
+    /**
+     * scheduleReconnect has been called
+     *
+     * @event websocket:scheduling-reconnect
+     * @param {Layer.Core.Event} evt
+     * @param {Number} evt.counter
+     * @param {Number} evt.delay
+     */
+    'websocket:scheduling-reconnect',
+
+    /**
+     * Websocket reconnect has been scheduled
+     *
+     * @event websocket:schedule-reconnect
+     * @param {Layer.Core.Event} evt
+     * @param {String} evt.from  Describes where the call to scheduleReconnect originated from
+     * @param {String} evt.why   Provides detail on why the call was made from there
+     */
+    'websocket:schedule-reconnect',
+
+    /**
+     * Websocket ignored skipped counters without checking for missed websocket data
+     *
+     * @event websocket:ignore-skipped-counter
+     * @param {Layer.Core.Event} evt
+     * @param {String} evt.from  Describes where the call to scheduleReconnect originated from
+     * @param {String} evt.why   Provides detail on why the call was made from there
+     */
+    'websocket:ignore-skipped-counter',
   ],
   lifecycle: {
 
@@ -38,6 +93,13 @@ const ClientWebsockets = {
       this.socketRequestManager = new WebsocketRequestManager({
         socketManager: this.socketManager,
       });
+
+      this.socketManager.on('connecting', evt => this.trigger('websocket:connecting'));
+      this.socketManager.on('disconnecting', evt => this.trigger('websocket:disconnecting', { from: evt.from, why: evt.why }));
+      this.socketManager.on('replaying-events', evt => this.trigger('websocket:replaying-events', { from: evt.from, why: evt.why }));
+      this.socketManager.on('schedule-reconnect', evt => this.trigger('websocket:schedule-reconnect', { from: evt.from, why: evt.why }));
+      this.socketManager.on('scheduling-reconnect', evt => this.trigger('websocket:scheduling-reconnect', { counter: evt.counter, delay: evt.delay }));
+      this.socketManager.on('ignore-skipped-counter', evt => this.trigger('websocket:ignore-skipped-counter', { from: evt.from, why: evt.why }));
     },
     destroy() {
       this.socketManager.destroy();